Bicycle Gan v2.0. Corrected the generator.

+++ loader.py
@@ -1,10 +1,7 @@
-import numpy as np
-import os
-import h5py
 import glob
-import scipy.misc
-from scipy.misc import imread, imresize
 
+import numpy as np
+from scipy.misc import imread
 
 def load_images(dataset_dir, option):
 	all_imgs = glob.glob(dataset_dir + "/" + option + "/" + "*.jpg")
@@ -18,8 +15,8 @@ def load_images(dataset_dir, option):
 		img_array_A.append(img_A)
 		img_array_B.append(img_B)
 
-	img_array_A = np.asarray(img_array_A) / 255
-	img_array_B = np.asarray(img_array_B) / 255
+	img_array_A = np.asarray(img_array_A).astype(np.float32) / 255
+	img_array_B = np.asarray(img_array_B).astype(np.float32) / 255
 	# print(train_A.shape)
 	#print(train_B.shape)
